implementing CICD and improving ml training

+++ dags/tasks/create_annoted_dataset.py
@@ -79,7 +79,7 @@ if __name__ == '__main__':
                              channel=args.channel,
                              start_date=args.start_time,
                              end_date=args.end_time,
-                             sampling_frequency_hz=float(
+                             sampling_frequency_hz=int(
                                  args.sampling_frequency_hz))
 
     df_ecg_annot = pd.concat([df_ecg, df_annot], axis=1).dropna()
